requirements.txt
python-dotenv
aiogram
openai
hugchat
python-docx